+++ src/internal/src/HRTFKernel.cpp
@@ -22,7 +22,7 @@ namespace lab
 // This represents the initial delay before the most energetic part of the impulse response.
 // The sample-frame delay is removed from the impulseP impulse response, and this value  is returned.
 // the length of the passed in AudioChannel must be a power of 2.
-inline float ExtractAverageGroupDelay(AudioChannel * channel, uint32_t analysisFFTSize)
+inline float ExtractAverageGroupDelay(AudioChannel * channel, size_t analysisFFTSize)
 {
     ASSERT(channel);
         
@@ -34,7 +34,7 @@ inline float ExtractAverageGroupDelay(AudioChannel * channel, uint32_t analysisF
         return 0;
     
     // Check for power-of-2.
-    ASSERT(1UL << static_cast<unsigned>(log2(analysisFFTSize)) == analysisFFTSize);
+    ASSERT(uint64_t(1) << static_cast<uint64_t>(log2(analysisFFTSize)) == analysisFFTSize);
 
     FFTFrame estimationFrame(analysisFFTSize);
     estimationFrame.doFFT(impulseP);
@@ -45,7 +45,8 @@ inline float ExtractAverageGroupDelay(AudioChannel * channel, uint32_t analysisF
     return frameDelay;
 }
 
-HRTFKernel::HRTFKernel(AudioChannel * channel, uint32_t fftSize, float sampleRate) : m_frameDelay(0) , m_sampleRate(sampleRate)
+HRTFKernel::HRTFKernel(AudioChannel * channel, size_t fftSize, float sampleRate) 
+: m_frameDelay(0) , m_sampleRate(sampleRate)
 {
     ASSERT(channel);
 
@@ -53,10 +54,10 @@ HRTFKernel::HRTFKernel(AudioChannel * channel, uint32_t fftSize, float sampleRat
     m_frameDelay = ExtractAverageGroupDelay(channel, fftSize / 2);
 
     float* impulseResponse = channel->mutableData();
-    uint32_t responseLength = channel->length();
+    size_t responseLength = channel->length();
 
     // We need to truncate to fit into 1/2 the FFT size (with zero padding) in order to do proper convolution.
-    uint32_t truncatedResponseLength = std::min(responseLength, fftSize / 2); // truncate if necessary to max impulse response length allowed by FFT
+    size_t truncatedResponseLength = std::min(responseLength, fftSize / 2); // truncate if necessary to max impulse response length allowed by FFT
 
     // Quick fade-out (apply window) at truncation point
     unsigned numberOfFadeOutFrames = static_cast<unsigned>(sampleRate / 4410); // 10 sample-frames @44.1KHz sample-rate
@@ -64,7 +65,7 @@ HRTFKernel::HRTFKernel(AudioChannel * channel, uint32_t fftSize, float sampleRat
 
     if (numberOfFadeOutFrames < truncatedResponseLength) 
     {
-        for (uint32_t i = truncatedResponseLength - numberOfFadeOutFrames; i < truncatedResponseLength; ++i) 
+        for (size_t i = truncatedResponseLength - numberOfFadeOutFrames; i < truncatedResponseLength; ++i) 
         {
             float x = 1.0f - static_cast<float>(i - (truncatedResponseLength - numberOfFadeOutFrames)) / numberOfFadeOutFrames;
             impulseResponse[i] *= x;
